requirements.txt
langchain>=0.3.0
langchain-core
langchain-postgres
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchain-openai>=0.3.0
openai>=1.50.0
pgvector>=0.3.6
psycopg[binary]>=3.2.0
python-dotenv>=1.0.1
pypdf>=5.0.0
